refactor: Extract federation event handling to an event processor
