+++ src/script/event/EventProcessor.ts
@@ -1,0 +1,30 @@
+/*
+ * Wire
+ * Copyright (C) 2023 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+import {BackendEvent} from '@wireapp/api-client/lib/event';
+
+import {EventSource} from './EventSource';
+
+import {ClientConversationEvent} from '../conversation/EventBuilder';
+
+export type IncomingEvent = BackendEvent | ClientConversationEvent;
+
+export interface EventProcessor {
+  processEvent(event: IncomingEvent, source: EventSource): Promise<void>;
+}
